운동 기록 게시글의 수정, 삭제 기능 구현

+++ app/src/main/java/frames/EditExerciseRecordPostFrame.java
@@ -8,6 +8,9 @@ import java.awt.*;
 import java.util.List;
 
 public class EditExerciseRecordPostFrame extends JFrame {
+  public static final int CREATION = 1;
+  public static final int MODIFICATION = 2;
+
   List<ExercisePlanPost> exercisePlanPosts;
   List<ExerciseRecordPost> exerciseRecordPosts;
   
@@ -19,9 +22,8 @@ public class EditExerciseRecordPostFrame extends JFrame {
   private ButtonGroup privacyStateCheckButtonGroup;
 
   public EditExerciseRecordPostFrame(
-      List<ExercisePlanPost> exercisePlanPosts,
-      ExercisePlanPost exercisePlanPost,
-      List<ExerciseRecordPost> exerciseRecordPosts) {
+      List<ExercisePlanPost> exercisePlanPosts, ExercisePlanPost exercisePlanPost,
+      List<ExerciseRecordPost> exerciseRecordPosts, int mode) {
     this.exercisePlanPosts = exercisePlanPosts;
     this.exerciseRecordPosts = exerciseRecordPosts;
 
@@ -36,11 +38,35 @@ public class EditExerciseRecordPostFrame extends JFrame {
     createStopoverPointsAchievementCheckPanel(exercisePlanPost.stopoverPoints());
     createExerciseDistanceAchievementsCheckPanel(exercisePlanPost.exerciseDistance());
     createFinalResultCheckPanel();
-    createDescriptionPanel();
+    createDescriptionPanel("");
     createPrivacyStateCheckPanel();
-    createButtonsPanel(exercisePlanPost);
+    createButtonsPanel(mode, exercisePlanPost, null);
 
     this.setVisible(true);
+  }
+
+  public EditExerciseRecordPostFrame(
+      List<ExerciseRecordPost> exerciseRecordPosts,
+      ExerciseRecordPost exerciseRecordPost, int mode) {
+    this.exerciseRecordPosts = exerciseRecordPosts;
+
+    this.setSize(500, 500);
+    this.setLocation(200, 150);
+    this.setLayout(new GridLayout(0, 1));
+
+    this.add(new JLabel(exerciseRecordPost.exercisePlanPost().title() + " 결과 기록하기"));
+    this.add(new JLabel("날짜: " + exerciseRecordPost.exercisePlanPost().date()));
+    this.add(new JLabel("운동 종류: " + exerciseRecordPost.exercisePlanPost().exerciseType()));
+    createExerciseTimeAchievementCheckPanel(exerciseRecordPost.exercisePlanPost().exerciseTime());
+    createStopoverPointsAchievementCheckPanel(exerciseRecordPost.exercisePlanPost().stopoverPoints());
+    createExerciseDistanceAchievementsCheckPanel(exerciseRecordPost.exercisePlanPost().exerciseDistance());
+    createFinalResultCheckPanel();
+    createDescriptionPanel(exerciseRecordPost.description());
+    createPrivacyStateCheckPanel();
+    createButtonsPanel(mode, null, exerciseRecordPost);
+
+    this.setVisible(true);
+
   }
 
   public void createExerciseTimeAchievementCheckPanel(String exerciseTime) {
@@ -103,10 +129,10 @@ public class EditExerciseRecordPostFrame extends JFrame {
     this.add(finalResultCheckPanel);
   }
 
-  public void createDescriptionPanel() {
+  public void createDescriptionPanel(String description) {
     JPanel descriptionPanel = new JPanel();
     descriptionPanel.add(new JLabel("상세 설명: "));
-    descriptionTextArea = new JTextArea();
+    descriptionTextArea = new JTextArea(description);
     descriptionTextArea.setColumns(20);
     descriptionTextArea.setRows(2);
     descriptionTextArea.setLineWrap(true);
@@ -121,18 +147,18 @@ public class EditExerciseRecordPostFrame extends JFrame {
     JPanel privacyStateCheckPanel = new JPanel();
     privacyStateCheckPanel.add(new JLabel("공개 여부: "));
     privacyStateCheckButtonGroup = new ButtonGroup();
-    JRadioButton achievedRadioButton = new JRadioButton("공개");
-    achievedRadioButton.setActionCommand("공개");
-    privacyStateCheckButtonGroup.add(achievedRadioButton);
-    privacyStateCheckPanel.add(achievedRadioButton);
-    JRadioButton notAchievedRadioButton = new JRadioButton("비공개");
-    notAchievedRadioButton.setActionCommand("비공개");
-    privacyStateCheckButtonGroup.add(notAchievedRadioButton);
-    privacyStateCheckPanel.add(notAchievedRadioButton);
+    JRadioButton setPublicRadioButton = new JRadioButton("공개");
+    setPublicRadioButton.setActionCommand("공개");
+    privacyStateCheckButtonGroup.add(setPublicRadioButton);
+    privacyStateCheckPanel.add(setPublicRadioButton);
+    JRadioButton setPrivateRadioButton = new JRadioButton("비공개");
+    setPrivateRadioButton.setActionCommand("비공개");
+    privacyStateCheckButtonGroup.add(setPrivateRadioButton);
+    privacyStateCheckPanel.add(setPrivateRadioButton);
     this.add(privacyStateCheckPanel);
   }
 
-  public void createButtonsPanel(ExercisePlanPost exercisePlanPost) {
+  public void createButtonsPanel(int mode, ExercisePlanPost exercisePlanPost, ExerciseRecordPost toBeModified) {
     JPanel buttonsPanel = new JPanel();
     buttonsPanel.setLayout(new BorderLayout());
     JButton cancelButton = new JButton("취소");
@@ -142,30 +168,58 @@ public class EditExerciseRecordPostFrame extends JFrame {
     buttonsPanel.add(cancelButton, BorderLayout.WEST);
     JButton registerButton = new JButton("등록하기");
     registerButton.addActionListener(event -> {
-      boolean achievedExerciseTime = exerciseTimeAchievementCheckButtonGroup
-          .getSelection().getActionCommand().equals("달성");
-      boolean visitedAllStopoverPoints = stopoverPointsAchievementCheckButtonGroup
-          .getSelection().getActionCommand().equals("달성");
-      boolean achievedExerciseDistance = exerciseDistanceAchievementCheckButtonGroup
-          .getSelection().getActionCommand().equals("달성");
-      boolean finalResult = finalResultCheckButtonGroup
-          .getSelection().getActionCommand().equals("달성");
-      String description = descriptionTextArea.getText();
-      boolean isPublicPost = privacyStateCheckButtonGroup
-          .getSelection().getActionCommand().equals("공개");
+      if (mode == EditExerciseRecordPostFrame.CREATION) {
+        boolean achievedExerciseTime = exerciseTimeAchievementCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        boolean visitedAllStopoverPoints = stopoverPointsAchievementCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        boolean achievedExerciseDistance = exerciseDistanceAchievementCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        boolean finalResult = finalResultCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        String description = descriptionTextArea.getText();
+        boolean isPublicPost = privacyStateCheckButtonGroup
+            .getSelection().getActionCommand().equals("공개");
 
-      ExerciseRecordPost exerciseRecordPost = new ExerciseRecordPost(
-          exercisePlanPost, achievedExerciseTime, visitedAllStopoverPoints,
-          achievedExerciseDistance, finalResult, description,
-          isPublicPost
-      );
+        ExerciseRecordPost exerciseRecordPost = new ExerciseRecordPost(
+            exercisePlanPost, achievedExerciseTime, visitedAllStopoverPoints,
+            achievedExerciseDistance, finalResult, description,
+            isPublicPost
+        );
 
-      exerciseRecordPosts.add(exerciseRecordPost);
+        exerciseRecordPosts.add(exerciseRecordPost);
 
-      for (ExercisePlanPost found : exercisePlanPosts) {
-        if (found.uniqueNumber() == exercisePlanPost.uniqueNumber()) {
-          found.delete();
-          break;
+        for (ExercisePlanPost found : exercisePlanPosts) {
+          if (found.uniqueNumber() == exercisePlanPost.uniqueNumber()) {
+            found.delete();
+            break;
+          }
+        }
+      }
+
+      if (mode == EditExerciseRecordPostFrame.MODIFICATION) {
+        boolean achievedExerciseTime = exerciseTimeAchievementCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        boolean visitedAllStopoverPoints = stopoverPointsAchievementCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        boolean achievedExerciseDistance = exerciseDistanceAchievementCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        boolean finalResult = finalResultCheckButtonGroup
+            .getSelection().getActionCommand().equals("달성");
+        String description = descriptionTextArea.getText();
+        boolean isPublicPost = privacyStateCheckButtonGroup
+            .getSelection().getActionCommand().equals("공개");
+
+        for (ExerciseRecordPost found : exerciseRecordPosts) {
+          if (found.uniqueNumber() == toBeModified.uniqueNumber()) {
+            found.modifyAchievedExerciseTime(achievedExerciseTime);
+            found.modifyVisitedAllStopoverPoints(visitedAllStopoverPoints);
+            found.modifyAchievedExerciseDistance(achievedExerciseDistance);
+            found.modifyFinalResult(finalResult);
+            found.modifyDescription(description);
+            found.modifyPrivacyOfPost(isPublicPost);
+            break;
+          }
         }
       }
